requirements.txt
Flask==2.3.2
openai==0.27.0
python-dotenv==0.21.0
requests==2.28.2